feature: implement email and Discord notification channels with secure options
