+++ internal/notifier/channels/default.go
@@ -1,4 +1,4 @@
-package notifier
+package channels
 
 import (
 	"fmt"
